fix(news): validate waste targeting payloads
