+++ packages/plugin-news/src/news.editor-model.ts
@@ -6,8 +6,11 @@ import type {
   NewsFormInput,
   NewsMediaContentFormValue,
   NewsSavePlan,
-  WasteLocationKey,
 } from './news.types.js';
+import {
+  deduplicateWasteLocationKeys,
+  mergeNewsWasteLocationKeys,
+} from './news.waste-payload.js';
 
 type NewsEditorialStatus = NewsSavePlan['editorialStatus'];
 type NewsLegacyCompatibilitySnapshot = NonNullable<
@@ -237,7 +240,7 @@ export const createNewsEditorFormValues = (
     },
     sourceUrlDescription: item.sourceUrl?.description ?? '',
     pushNotificationEnabled: false,
-    wasteLocationKeys: deduplicateWasteLocationKeys(item.payload.wasteLocationKeys ?? []),
+    wasteLocationKeys: deduplicateWasteLocationKeys(item.payload.wasteLocationKeys),
     publicationMode:
       editorialStatus === 'draft'
         ? 'draft'
@@ -285,67 +288,4 @@ export const buildNewsSavePayload = (
       ...(payload ? { payload } : {}),
     },
   };
-};
-
-const normalizeWasteLocationKey = (key: WasteLocationKey): WasteLocationKey => ({
-  street: key.street.trim(),
-  zip: key.zip.trim(),
-  city: key.city.trim(),
-});
-
-const wasteLocationKeyIdentity = (key: WasteLocationKey): string =>
-  JSON.stringify([key.street, key.zip, key.city]);
-
-export const deduplicateWasteLocationKeys = (
-  keys: readonly WasteLocationKey[]
-): WasteLocationKey[] => {
-  const result = new Map<string, WasteLocationKey>();
-  for (const value of keys) {
-    const key = normalizeWasteLocationKey(value);
-    if (!key.street || !key.zip || !key.city) continue;
-    result.set(wasteLocationKeyIdentity(key), key);
-  }
-  return [...result.values()];
-};
-
-export const mergeNewsWasteLocationKeys = (
-  existingPayload: NewsContentItem['payload'] | undefined,
-  keys: readonly WasteLocationKey[]
-): NewsContentItem['payload'] | undefined => {
-  const nextPayload = { ...(existingPayload ?? {}) };
-  const normalizedKeys = deduplicateWasteLocationKeys(keys);
-  const removesExistingWasteLocationKeys =
-    existingPayload !== undefined &&
-    'wasteLocationKeys' in existingPayload &&
-    normalizedKeys.length === 0;
-  if (normalizedKeys.length === 0) {
-    delete nextPayload.wasteLocationKeys;
-  } else {
-    nextPayload.wasteLocationKeys = normalizedKeys;
-  }
-  return Object.keys(nextPayload).length > 0 || removesExistingWasteLocationKeys
-    ? nextPayload
-    : undefined;
-};
-
-export const requiresGlobalPushConfirmation = (input: {
-  readonly pushNotificationEnabled: boolean;
-  readonly targetCount: number;
-  readonly pushNotificationsSentAt?: string;
-}): boolean =>
-  input.pushNotificationEnabled &&
-  input.targetCount === 0 &&
-  !input.pushNotificationsSentAt;
-
-export type WasteTargetingAvailability = 'available' | 'forbidden' | 'load-error' | 'loading';
-
-export const resolveGlobalPushConfirmationKey = (
-  availability: WasteTargetingAvailability
-):
-  | 'targeting.globalConfirm.noTargets'
-  | 'targeting.globalConfirm.forbidden'
-  | 'targeting.globalConfirm.loadError' => {
-  if (availability === 'available') return 'targeting.globalConfirm.noTargets';
-  if (availability === 'forbidden') return 'targeting.globalConfirm.forbidden';
-  return 'targeting.globalConfirm.loadError';
 };
